[#161184689] Fix elasticsearch absolute url to repository files.
